Moved some functions based on the most recent feedback on PR
