+++ app/logic/allPendingForms.py
@@ -16,7 +16,6 @@ from app.models.notes import Notes
 from app.login_manager import DoesNotExist, render_template
 
 
-
 def saveStatus(new_status, form_ids, currentUser):
     try:
         if new_status == 'Denied by Admin':
@@ -254,3 +253,36 @@ def financialAidSAASOverloadApproval(historyForm, rsp, status, currentUser, curr
         selectedOverload.SAASReviewDate = currentDate
     selectedOverload.save()
     return jsonify({"Success": True})
+
+
+def checkAdjustment(allForms):
+    """
+        Retrieve supervisor and position information for adjusted forms using the new values
+        stored in adjusted table and update allForms
+    """
+    if allForms.adjustedForm:
+
+        if allForms.adjustedForm.fieldAdjusted == "supervisor":
+            # use the supervisor id in the field adjusted to find supervisor in User table.
+            newSupervisorID = allForms.adjustedForm.newValue
+            newSupervisor = Supervisor.get(Supervisor.ID == newSupervisorID)
+            if not newSupervisor:
+                newSupervisor = createSupervisorFromTracy(bnumber=newSupervisorID)
+
+            # we are temporarily storing the supervisor name in new value,
+            # because we want to show the supervisor name in the hmtl template.
+            allForms.adjustedForm.newValue = newSupervisor.FIRST_NAME +" "+ newSupervisor.LAST_NAME
+            allForms.adjustedForm.oldValue = {"email":newSupervisor.EMAIL, "ID":newSupervisor.ID}
+
+        if allForms.adjustedForm.fieldAdjusted == "position":
+            newPositionCode = allForms.adjustedForm.newValue
+            newPosition = Tracy().getPositionFromCode(newPositionCode)
+            # temporarily storing the position code and wls in new value, and position name in old value
+            # because we want to show these information in the hmtl template.
+            allForms.adjustedForm.newValue = newPosition.POSN_CODE +" (" + newPosition.WLS+")"
+            allForms.adjustedForm.oldValue = newPosition.POSN_TITLE
+
+        if allForms.adjustedForm.fieldAdjusted == "department":
+            newDepartment = Department.get(Department.ORG==allForms.adjustedForm.newValue)
+            allForms.adjustedForm.newValue = newDepartment.DEPT_NAME
+            allForms.adjustedForm.oldValue = newDepartment.ORG + "-" + newDepartment.ACCOUNT
