--- app/logic/studentsFromTracy.py
@@ -1,9 +1,0 @@
-from app.logic.tracy import Tracy
-from app.logic.tracy import InvalidQueryException
-from peewee import DoesNotExist
-from app.models.student import Student
-from app.models.user import User
-
-
-
-
