--- app/logic/user.py
@@ -1,1 +1,0 @@
-# To be removed
