+++ app/logic/userInsertFunctions.py
@@ -12,7 +12,7 @@ from app.models.term import *
 from app.models.student import Student
 from app.models.supervisor import Supervisor
 from app.models.department import *
-from app.logic.tracy import Tracy, InvalidQueryException, createStudentFromTracy, InvalidUserException
+from app.logic.tracy import Tracy, InvalidQueryException, InvalidUserException
 
 
 def updatePersonRecords():
@@ -106,44 +106,6 @@ def updateSupervisorRecord(supervisor):
 
 
 
-
-
-
-def createSupervisorFromTracy(username=None, bnumber=None):
-    """
-        Attempts to add a supervisor from the Tracy database to the application, based on the provided username or bnumber.
-
-        Raises InvalidUserException if this does not succeed.
-    """
-    if bnumber:
-        try:
-            tracyUser = Tracy().getSupervisorFromID(bnumber)
-        except InvalidQueryException as e:
-            raise InvalidUserException("{} not found in Tracy database".format(bnumber))
-
-    else:    # Executes if no ID is provided
-        email = "{}@berea.edu".format(username)
-        try:
-            tracyUser = Tracy().getSupervisorFromEmail(email)
-        except InvalidQueryException as e:
-            raise InvalidUserException("{} not found in Tracy database".format(email))
-
-    try:
-        return Supervisor.get(Supervisor.ID == tracyUser.ID.strip())
-    except DoesNotExist:
-        return Supervisor.create(PIDM = tracyUser.PIDM,
-                                 legal_name = tracyUser.FIRST_NAME,
-                                 LAST_NAME = tracyUser.LAST_NAME,
-                                 ID = tracyUser.ID.strip(),
-                                 EMAIL = tracyUser.EMAIL,
-                                 CPO = tracyUser.CPO,
-                                 ORG = tracyUser.ORG,
-                                 DEPT_NAME = tracyUser.DEPT_NAME)
-    except Exception as e:
-        print(e)
-        raise InvalidUserException("Error: Could not get or create {0} {1}".format(tracyUser.FIRST_NAME, tracyUser.LAST_NAME))
-
-
 def createUser(username, student=None, supervisor=None):
     """
     Retrieves or creates a user in the User table and updates Supervisor and/or Student as requested.
@@ -184,3 +146,83 @@ def getOrCreateStudentRecord(username=None, bnumber=None):
     except DoesNotExist:
         student = createStudentFromTracy(username,bnumber)
     return student
+
+
+
+def createStudentFromTracy(username=None, bnumber=None):
+    """
+        Attempts to add a student from the Tracy database to the application, based on the provided username or bnumber.
+
+        Raises InvalidUserException if this does not succeed.
+    """
+    if not username and not bnumber:
+        raise ValueError("No arguments provided to createStudentFromTracy()")
+
+    if bnumber:
+        try:
+            tracyStudent = Tracy().getStudentFromBNumber(bnumber)
+        except InvalidQueryException as e:
+            raise InvalidUserException("{} not found in Tracy database".format(bnumber))
+
+    else:    # Executes if no ID is provided
+        email = "{}@berea.edu".format(username)
+        try:
+            tracyStudent = Tracy().getStudentFromEmail(email)
+        except InvalidQueryException as e:
+            raise InvalidUserException("{} not found in Tracy database".format(email))
+
+    # Create the student in Tracy
+    try:
+        return Student.get(Student.ID == tracyStudent.ID.strip())
+    except DoesNotExist:
+        #print('Could not find {0} {1} in Student table, creating new entry.'.format(tracyStudent.FIRST_NAME, tracyStudent.LAST_NAME))
+        return Student.create(ID = tracyStudent.ID.strip(),
+                            PIDM = tracyStudent.PIDM,
+                            legal_name = tracyStudent.FIRST_NAME,
+                            LAST_NAME = tracyStudent.LAST_NAME,
+                            CLASS_LEVEL = tracyStudent.CLASS_LEVEL,
+                            ACADEMIC_FOCUS = tracyStudent.ACADEMIC_FOCUS,
+                            MAJOR = tracyStudent.MAJOR,
+                            PROBATION = tracyStudent.PROBATION,
+                            ADVISOR = tracyStudent.ADVISOR,
+                            STU_EMAIL = tracyStudent.STU_EMAIL,
+                            STU_CPO = tracyStudent.STU_CPO,
+                            LAST_POSN = tracyStudent.LAST_POSN,
+                            LAST_SUP_PIDM = tracyStudent.LAST_SUP_PIDM)
+    else:
+        raise InvalidUserException("Error: Could not get or create {0} {1}".format(tracyStudent.FIRST_NAME, tracyStudent.LAST_NAME))
+
+
+def createSupervisorFromTracy(username=None, bnumber=None):
+    """
+        Attempts to add a supervisor from the Tracy database to the application, based on the provided username or bnumber.
+
+        Raises InvalidUserException if this does not succeed.
+    """
+    if bnumber:
+        try:
+            tracyUser = Tracy().getSupervisorFromID(bnumber)
+        except InvalidQueryException as e:
+            raise InvalidUserException("{} not found in Tracy database".format(bnumber))
+
+    else:    # Executes if no ID is provided
+        email = "{}@berea.edu".format(username)
+        try:
+            tracyUser = Tracy().getSupervisorFromEmail(email)
+        except InvalidQueryException as e:
+            raise InvalidUserException("{} not found in Tracy database".format(email))
+
+    try:
+        return Supervisor.get(Supervisor.ID == tracyUser.ID.strip())
+    except DoesNotExist:
+        return Supervisor.create(PIDM = tracyUser.PIDM,
+                                 legal_name = tracyUser.FIRST_NAME,
+                                 LAST_NAME = tracyUser.LAST_NAME,
+                                 ID = tracyUser.ID.strip(),
+                                 EMAIL = tracyUser.EMAIL,
+                                 CPO = tracyUser.CPO,
+                                 ORG = tracyUser.ORG,
+                                 DEPT_NAME = tracyUser.DEPT_NAME)
+    except Exception as e:
+        print(e)
+        raise InvalidUserException("Error: Could not get or create {0} {1}".format(tracyUser.FIRST_NAME, tracyUser.LAST_NAME))
